+++ src/exchange.ts
@@ -374,7 +374,7 @@ export class Exchange {
     if (loadConfig.loadAssets) {
       this._assets = loadConfig.loadAssets;
     } else {
-      this._assets = assets.allAssets();
+      this._assets = assets.allAssets(loadConfig.network);
     }
     this._provider = new anchor.AnchorProvider(
       loadConfig.connection,